% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacseq.R
\name{atac_report}
\alias{atac_report}
\title{atac_report_single}
\usage{
atac_report(input, output)
}
\arguments{
\item{input}{directory to the sample}

\item{output}{directory to the html file}

\item{template}{the template, default from hiseqr}
}
\description{
atac_report_single
}
