% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffbind.R
\name{atacseq_diff}
\alias{atacseq_diff}
\title{atacseq_diff}
\usage{
atacseq_diff(dirA, dirB, outdir = NULL)
}
\arguments{
\item{dirA}{path to directory of merge sample, B}

\item{dirB}{path to directory of merge sample, B}

\item{outdir}{path to the output, dirA.vs.dirB}
}
\description{
atacseq_diff
}
