% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{bacteria}
\alias{bacteria}
\alias{hiseq_kraken2_report}
\title{Functions for bacteria}
\usage{
hiseq_kraken2_report(indir, outdir, preview = FALSE)
}
\arguments{
\item{indir}{character Path to the of kraken2 results}

\item{outdir}{character path to the html report}
}
\description{
Create html report for kraken2 output
}
\examples{
\donotrun{

}

}
