% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Functions for HiSeq plots}
\usage{
bar_plot(
  data = NULL,
  x,
  y,
  label = y,
  group = NULL,
  fill = group,
  direction = "vertical",
  position = "stack",
  ...
)
}
\arguments{
\item{data}{data.frame require id, count column}

\item{x}{character Column on x-axis}

\item{y}{character Column on y-axis}

\item{direction}{character Options: vertical, horizontal, default: vertical}

\item{...}{extra arguments for plots}
}
\description{
Including bar_plot, scatter_plot, venn_plot, ma_plot,  box_plot, ...
}
\details{
bar_plot Make barplot
}
