% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_plot_lendist}
\alias{bar_plot_lendist}
\title{length distribution}
\usage{
bar_plot_lendist(df, x = "length", y = "count", ...)
}
\arguments{
\item{...}{optional arguments
range 18:40
color_list c("red", "blue")}
}
\description{
length distribution
}
\section{Functions}{
\itemize{
\item \code{bar_plot_lendist()}: for length distribution

}}
