% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{bed_venn}
\alias{bed_venn}
\title{bed_venn}
\usage{
bed_venn(blist, names = NULL)
}
\arguments{
\item{blist}{list, a list of BED files}

\item{names}{character, the names for each group}
}
\description{
bed_venn
}
