% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_cols2}
\alias{bind_cols2}
\title{bind_cols2}
\usage{
bind_cols2(..., by = NULL)
}
\arguments{
\item{by}{column name for function merge()

see example on https://www.r-bloggers.com/2018/10/how-to-perform-merges-joins-on-two-or-more-data-frames-with-base-r-tidyverse-and-data-table/
see dplyr::bind_cols, for how to process dots}

\item{x}{a list of data.frames}
}
\description{
bind_cols2
}
