% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_hash_string}
\alias{check_hash_string}
\title{check_hash_string}
\usage{
check_hash_string(hash_x, x)
}
\arguments{
\item{hash_x}{string, SHA-256 value}

\item{x}{string}
}
\description{
check_hash_string
}
