% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{chipseq_report}
\alias{chipseq_report}
\title{chipseq_report_single}
\usage{
chipseq_report(input, output)
}
\arguments{
\item{input}{directory to the sample}

\item{output}{directory to the html file}

\item{template}{the template, default from hiseqr}
}
\description{
chipseq_report_single
}
