% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors}
\alias{colors}
\alias{gg_color}
\title{Functions for colors}
\usage{
gg_color(n = 3)
}
\arguments{
\item{n}{integer}
}
\description{
This is an implementation of pick_colors, from ggplot2 default, or
RColorBrewer, fishualize
}
