% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{convert_id}
\alias{convert_id}
\title{convert_id}
\usage{
convert_id(
  x,
  from_keytype = NULL,
  to_keytype = "SYMBOL",
  organism = NULL,
  orgdb = NULL,
  rm_na = FALSE,
  ...
)
}
\arguments{
\item{x}{gene names}

\item{organism}{name of the genome, eg: "dm6"}
}
\description{
convert_id
}
