% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_report.R
\name{csv_to_html_report}
\alias{csv_to_html_report}
\title{csv_to_html}
\usage{
csv_to_html_report(input, output)
}
\arguments{
\item{input}{csv file}

\item{output}{html file}

\item{template}{the template, default from hiseqr}
}
\description{
csv_to_html
}
