% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{deseq2_main}
\alias{deseq2_main}
\title{deseq2_main}
\usage{
deseq2_main(
  dds,
  outdir,
  fc_cutoff = 2,
  pval_cutoff = 0.05,
  organism = NULL,
  readable = TRUE
)
}
\arguments{
\item{dds}{count data in matrix}

\item{outdir}{Directory to save the results}

\item{fc_cutoff}{float Cutoff for foldchange, default: 2}

\item{pval_cutoff}{float Cutoff for pvalue, default: 0.05}

\item{organism}{character, the name of the organism, eg: dm6, hg38, human,}

\item{readable}{bool, Add symbol and entrezid to table}
}
\description{
deseq2_main
}
