% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{deseq_cmp
functions for pair deseq}
\alias{deseq_cmp
functions for pair deseq}
\alias{deseq_pair_report}
\title{Functions for RNAseq compare analysis}
\usage{
deseq_pair_report(x, y, outdir, shrink_method = "apeglm")
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{outdir}{string directory to save files}

\item{feature}{string default: gene}
}
\description{
Apply DESeq2 for RNAseq DE analysis
require >=2 replicates for each sample
}
\details{
input: count.txt, from featureCounts output
output: DESeq2 output

helper functions:
}
