% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{deseq_cmp}
\alias{deseq_cmp}
\title{Functions for RNAseq compare analysis}
\usage{
deseq_cmp(x)
}
\description{
A tmp function, help documentation
}
\details{
Apply DESeq2 for RNAseq DE analysis
require >=2 replicates for each sample

input: count.txt, from featureCounts output
output: DESeq2 output

helper functions:

deseq_cmp
}
