% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{deseq_mean}
\alias{deseq_mean}
\title{deseq_mean}
\usage{
deseq_mean(x, ...)
}
\arguments{
\item{x}{DESeqDataSet, or '.csv' parsing design from \code{colData(dds)}}

\item{outdir}{for run_deseq_res()}

\item{transform_method}{character for dds count transformed,
'standard', 'vst', 'rlog'; default: 'standard'}
}
\value{
data.frame
}
\description{
deseq_mean
}
