% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{deseq_pair_alluvial_plot}
\alias{deseq_pair_alluvial_plot}
\title{RNAseq cmp,
number of sig genes}
\usage{
deseq_pair_alluvial_plot(x = NULL, y = NULL, feature = "gene", .pd_rds = NULL)
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{feature}{string default: gene}

\item{.pd_rds}{string path to the file, saving read_deseq_pair() output}
}
\description{
RNAseq cmp,
number of sig genes
}
