% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{deseq_pair_heatmap_te}
\alias{deseq_pair_heatmap_te}
\title{deseq_pair_heatmap_te}
\usage{
deseq_pair_heatmap_te(
  x = NULL,
  y = NULL,
  feature = "te",
  topN = 70,
  .pd_rds = NULL
)
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{feature}{string default: gene}

\item{.pd_rds}{string path to the file, saving read_deseq_pair() output}
}
\description{
deseq_pair_heatmap_te
}
