% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{deseq_pair_report}
\alias{deseq_pair_report}
\title{deseq_pair_report}
\usage{
deseq_pair_report(x, y, outdir, shrink_method = "apeglm")
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{outdir}{string directory to save files}

\item{feature}{string default: gene}
}
\description{
deseq_pair_report
}
