% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_plot.R
\name{deseq_qc_add_sig_label}
\alias{deseq_qc_add_sig_label}
\title{deseq_qc_add_sig_label}
\usage{
deseq_qc_add_sig_label(x, ...)
}
\arguments{
\item{x}{ggplot}

\item{label_list}{character gene names
priority: symbol > gene_id}
}
\value{
ggplot
}
\description{
deseq_qc_add_sig_label
}
