% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_plot.R
\name{deseq_qc_dist}
\alias{deseq_qc_dist}
\title{deseq_qc_dist}
\usage{
deseq_qc_dist(x, ...)
}
\arguments{
\item{x}{DESeqDataSet}

\item{outdir}{character saving the results}

\item{transform_method}{bool compute tranformed data, \code{vst()}, \code{rlog()}
default: standard}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
ggplot
}
\description{
deseq_qc_dist
}
