% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_plot.R
\name{deseq_qc_ma}
\alias{deseq_qc_ma}
\title{deseq_qc_ma}
\usage{
deseq_qc_ma(x, ...)
}
\arguments{
\item{x}{see \code{deseq_qc_res(x=)}}

\item{outdir}{character saving the results}

\item{ylim}{numeric, limits for log2FoldChange, default: NULL,
calculate by \code{scales::breaks_ext(n=5)(df$log2FoldChange)}}

\item{fc}{numeric cutoff for foldchange, default: 1, ignore foldchange}

\item{pvalue}{numeric cutoff for padj, default: 0.1, the main criteria}

\item{p_adjust}{bool use p-adjust value instead}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
ggplot
}
\description{
deseq_qc_ma
}
