% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_plot.R
\name{deseq_qc_mean_sd}
\alias{deseq_qc_mean_sd}
\title{deseq_qc_mean_sd
check mean standard diviation (SD), after transformation by vst, vlog}
\usage{
deseq_qc_mean_sd(x, ...)
}
\arguments{
\item{x}{DESeqDataSet}

\item{outdir}{character saving the results}

\item{transform_method}{bool compute tranformed data, \code{vst()}, \code{rlog()}
default: standard}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
ggplot
}
\description{
deseq_qc_mean_sd
check mean standard diviation (SD), after transformation by vst, vlog
}
