% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_plot.R
\name{deseq_qc_pca}
\alias{deseq_qc_pca}
\title{deseq_qc_pca}
\usage{
deseq_qc_pca(x, ...)
}
\arguments{
\item{x}{\code{DESeqDataSet}, see \code{deseq_qc_dds(x=)}}

\item{outdir}{character saving the results}

\item{transform}{bool compute tranformed data, \code{vst()}, \code{rlog()}
default: FALSE}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
ggplot
}
\description{
deseq_qc_pca
}
