% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_plot.R
\name{deseq_qc_plot}
\alias{deseq_qc_plot}
\alias{deseq_qc_counts}
\title{Functions for deseq quality control, utils}
\usage{
deseq_qc_counts(x, ...)
}
\arguments{
\item{x}{DESeqDataSet}

\item{outdir}{character saving the results}

\item{transform_method}{bool compute tranformed data, \code{vst()}, \code{rlog()}
default: standard}

\item{n_max}{integer number of genes to display, defualt: 16}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
ggplot
}
\description{
deseq_qc_dds(), parsing dds data from deseq dir
deseq_qc_res(), parsing res data from deseq dir
deseq_qc_counts()
deseq_qc_mean_sd()
deseq_qc_top_gene()
deseq_qc_dist()
deseq_qc_pca()
deseq_qc_ma()
deseq_qc_volcano()
deseq_qc_scatter()
deseq_qc_add_sig_label() add labels for ma, volcano, scatter
}
\details{
to-to: simpify arguments by \code{...}
}
