% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_utils.R
\name{deseq_qc_res}
\alias{deseq_qc_res}
\title{deseq_qc_res}
\usage{
deseq_qc_res(x, ...)
}
\arguments{
\item{x}{character path to the file \code{norm_table.fix.xls}
also support data types: \code{DESeqDataSet}, ...}

\item{log2fc_limits}{numeric setting the range of log2FoldChange on plots
default: NULL, by \code{scales::breaks_extend()}}

\item{fc}{numeric cutoff for foldchange, default: 1, ignore foldchange}

\item{pvalue}{numeric cutoff for padj, default: 0.1, the main criteria}

\item{p_adjust}{bool use p-adjust value instead}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
ggplot
}
\description{
deseq_qc_res
}
