% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_plot.R
\name{deseq_qc_scatter}
\alias{deseq_qc_scatter}
\title{deseq_qc_scatter}
\usage{
deseq_qc_scatter(x, ...)
}
\arguments{
\item{x}{see \code{deseq_qc_res(x=)}}

\item{outdir}{character saving the results}

\item{fc}{numeric cutoff for foldchange, default: 1, ignore foldchange}

\item{pvalue}{numeric cutoff for padj, default: 0.1, the main criteria}

\item{p_adjust}{bool use p-adjust value instead}

\item{density_points}{bool use \code{stat_density_2d()}, for
large number dots}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
ggplot
}
\description{
deseq_qc_scatter
}
