% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_plot.R
\name{deseq_qc_top_gene}
\alias{deseq_qc_top_gene}
\title{deseq_qc_top_gene}
\usage{
deseq_qc_top_gene(x, ...)
}
\arguments{
\item{x}{\code{DESeqDataSet}, see \code{deseq_qc_dds(x=)}}

\item{outdir}{character saving the results}

\item{transform_method}{bool compute tranformed data, \code{vst()}, \code{rlog()}
default: standard}

\item{n_max}{integer number of genes to display, defualt: 30}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
ggplot
}
\description{
deseq_qc_top_gene
}
