% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc_utils.R
\name{deseq_qc_utils}
\alias{deseq_qc_utils}
\alias{deseq_qc_dds}
\title{Functions for deseq quality control, utils}
\usage{
deseq_qc_dds(x = NULL, ...)
}
\arguments{
\item{x}{DESeqDataSet object or outdir (deseq_deseq2)}
}
\value{
list of standard*, vst, rlog objects
}
\description{
deseq_qc_dds(), parsing dds data from deseq dir
deseq_qc_res(), parsing res data from deseq dir
deseq_qc_counts()
deseq_qc_mean_sd()
deseq_qc_top_gene()
deseq_qc_dist()
deseq_qc_pca()
deseq_qc_ma()
deseq_qc_volcano()
deseq_qc_scatter()
deseq_qc_add_sig_label() add labels for ma, volcano, scatter
}
\details{
to-to: simpify arguments by \code{...}
}
