% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{deseq_sanitize_str}
\alias{deseq_sanitize_str}
\title{sanitize_str}
\usage{
deseq_sanitize_str(x, n_max = 0, fix_prefix = FALSE)
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
sanitize_str
}
