% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffbind.R
\name{diffbind
get_hiseq_data

retrive data from hiseq directory}
\alias{diffbind
get_hiseq_data

retrive data from hiseq directory}
\alias{get_hiseq_data}
\title{Functions for DiffBind analysis}
\usage{
get_hiseq_data(dir, group = "peak")
}
\arguments{
\item{dir}{path to the hiseq}

\item{group}{attribute of the file, peak, bam, ...}
}
\description{
See documentation: http://bioconductor.org/packages/release/bioc/vignettes/DiffBind/inst/doc/DiffBind.pdf
}
\details{
! to-do
peak annotation
Support: dm6, only (up-to-date)

Input design, outdir
Output diff peaks, plots

helper functions:
}
