% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffbind.R
\name{diffbind_hub}
\alias{diffbind_hub}
\title{Run DiffBind for peaks (groups)}
\usage{
diffbind_hub(design, outdir)
}
\arguments{
\item{design}{data.frame or csv file, for the peak and bam files}

\item{outdir}{output dirs}
}
\description{
Standard
}
