% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.hiseq_plots2.R, R/hiseq_plots.R
\name{.fix_prefix}
\alias{.fix_prefix}
\title{remove the longest common prefix from strings}
\usage{
.fix_prefix(x, replacement = "r")

.fix_prefix(x, replacement = "r")
}
\description{
RNAseq_wt_rep1, RNAseq_wt_rep2 => r1, r2

RNAseq_wt_rep1, RNAseq_wt_rep2 => r1, r2
}
