% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_report.R
\name{.url_to_link}
\alias{.url_to_link}
\title{Convert url to link in html
resize the image by scale
\if{html}{\out{<img src="img.jpg" alt="a img" style="width:500px;height:600px;">}}}
\usage{
.url_to_link(x, ...)
}
\description{
Convert url to link in html
resize the image by scale
\if{html}{\out{<img src="img.jpg" alt="a img" style="width:500px;height:600px;">}}
}
