% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falco.R
\name{falco_fetch_data}
\alias{falco_fetch_data}
\title{get_falco_data}
\usage{
falco_fetch_data(x, module = "ALL", group = "ALL")
}
\arguments{
\item{x}{full list of falco_read_txt}
}
\description{
get_falco_data
}
