% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falco.R
\name{falco_plot_lendist}
\alias{falco_plot_lendist}
\title{falco_plot_lendist
bar plot or line plot
module = 7, Sequence Length Distribution}
\usage{
falco_plot_lendist(x, mode = "line", ...)
}
\arguments{
\item{x}{output of falco_read_txt()}

\item{mode}{character line, bar}
}
\description{
falco_plot_lendist
bar plot or line plot
module = 7, Sequence Length Distribution
}
