% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falco.R
\name{falco_read_txt}
\alias{falco_read_txt}
\title{read_falco}
\usage{
falco_read_txt(x, module = "ALL", group = "ALL")
}
\arguments{
\item{x}{string fastqc_data.txt file}

\item{module}{int or character the name of the modules}

\item{group}{character data, id, flag, extra}
}
\description{
read_falco
}
