% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.fastqc.R
\name{fastqc_files}
\alias{fastqc_files}
\title{fastqcFiles}
\usage{
fastqc_files(x)

fastqc_files(x)
}
\arguments{
\item{x}{Path to the input directory}
}
\value{
A list of paths of zip files

A list of paths of zip files
}
\description{
This function parsing *_fastqc.zip files from input path.
It assumes that input path is a directory or a single *_fastqc.zip
file.

This function parsing *_fastqc.zip files from input path.
It assumes that input path is a directory or a single *_fastqc.zip
file.
}
