% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.fastqc.R
\name{fastqc_report}
\alias{fastqc_report}
\title{fastqc_report}
\usage{
fastqc_report(indir, outdir, preview = FALSE)
}
\arguments{
\item{indir}{Path to the input directory, Default is the
current working directory.}

\item{outdir}{path to the result file prefix (e.g., path/to/qc-result).
Don't add the file extension.}

\item{preview}{logical value. If TRUE, shows a preview of the report.}

\item{template}{a character vector specifying the path to an Rmd template.
file.}
}
\value{
A list of paths of zip files
}
\description{
This function parsing *_fastqc.zip files from input path.
It assumes that input path is a directory or a single *_fastqc.zip
file.
}
\details{
Create html report for fastqc output
}
