% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{filt_sig_gene}
\alias{filt_sig_gene}
\title{filt_sig_gene}
\usage{
filt_sig_gene(x, type = "sig", ...)
}
\arguments{
\item{x}{data.frame, DESeqResults, matrix}

\item{type}{character could be combination of
\link{"all", "sig", "up", "down", "not"},
default: "sig"}

\item{fc}{numeric, cutoff for foldchange, default: 2}

\item{pvalue}{numeric, cutoff for pvalue, default: 0.05}

\item{p_adjust}{bool use p-adjust value instead}

\item{force}{logical force calculate sig, default: FALSE}
}
\description{
filt_sig_gene
}
