% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{fix_hiseq_lib_type}
\alias{fix_hiseq_lib_type}
\title{fix_hiseq_lib_type}
\usage{
fix_hiseq_lib_type(x)
}
\arguments{
\item{x}{character the filename or lib_type}
}
\description{
in the filename, HiSeq located in the first part
Standard hiseq names in this package

RNAseq
ChIPseq
ATACseq
DNAseq
CnR (CUT&RUN)
CnT (CUT&TAG)
smRNAseq (small RNAseq)
STACCseq
HiC
RiboSeq
GoldCLIP
CLIPseq
RIPseq
GridSeq
}
