% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{fix_hiseq_lib_user}
\alias{fix_hiseq_lib_user}
\title{lib_number, auto_fix
eg: YY01, WM03}
\usage{
fix_hiseq_lib_user(x)
}
\arguments{
\item{x}{character lib_user}
}
\description{
lib_number, auto_fix
eg: YY01, WM03
}
