% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_report.R
\name{fix_hiseq_names}
\alias{fix_hiseq_names}
\title{fix samples names}
\usage{
fix_hiseq_names(x, ...)
}
\arguments{
\item{x}{names}
}
\description{
fix hiseq sample names, remove the most common string in names
}
