% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{fix_hiseq_sample_name}
\alias{fix_hiseq_sample_name}
\title{mission:
\enumerate{
\item replace non-letters, by "_"
\item fix hiseq_type
\item add rep1/rep2 in the tail
}}
\usage{
fix_hiseq_sample_name(x)
}
\arguments{
\item{x}{character}
}
\description{
mission:
\enumerate{
\item replace non-letters, by "_"
\item fix hiseq_type
\item add rep1/rep2 in the tail
}
}
