% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fq_name}
\alias{fq_name}
\title{fq_name}
\usage{
fq_name(x, fix_pe = TRUE, fix_rep = FALSE)
}
\arguments{
\item{x}{character, name of fastq files}

\item{fix_pe}{bool, trim _1, _2 suffix, default: FALSE}

\item{fix_rep}{bool, trim _rep1, _r1 in tha file name}
}
\description{
fq_name
}
