% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fragsize_plot}
\alias{fragsize_plot}
\title{fragsize_plot line plot for fragment size
for x-range}
\usage{
fragsize_plot(
  data,
  xmin = 0,
  xmax = 500,
  log10_y = FALSE,
  position = "fill",
  ...
)
}
\arguments{
\item{data}{data.frame for plotting}

\item{xmin}{int Min of the fragsize, default: 0}

\item{xmax}{int Max of the fragsize, defualt: 1000}

\item{log10_y}{bool Convert y-axis to log10 scale}

\item{position}{character, fill percentage, others show the count}

\item{...}{optional arguments
log10_y = FALSE
position = "fill"
= 18:40,
color_list = c("red", "blue")}
}
\description{
fragsize_plot line plot for fragment size
for x-range
}
