% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{gene_to_link}
\alias{gene_to_link}
\title{gene_to_link}
\usage{
gene_to_link(
  x,
  organism,
  style = "url",
  site = "https://asia.ensembl.org",
  readable = FALSE
)
}
\arguments{
\item{x}{string, gene names}

\item{organism}{string, scientific name of the organism, eg: Drosophila melanogaster,}

\item{style}{string, url, markdown, html, ...}
}
\description{
create emsembl geneID links
format \link{site}/id/\link{stable_id}
site1: https://asia.ensembl.org
site2: https://www.ensembl.org
ensembl: https://asia.ensembl.org/id/FBgn0004872
flybase: http://flybase.org/reports/FBgn0004872
}
