% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get_fix_csv}
\alias{get_fix_csv}
\title{#' @describeIn get_fix_csv Parsing the fix.xls file from DESeq2
#'
#'
#' @export
get_fix_csv <- function(x) {
if(is.character(x)) {
x <- x\link{1}
} else {
on.exit(message("'x' require character"), add = TRUE)
return(NULL)
}RNAseqRx directoryx/deseq/transcripts_deseq2.fix.xls

hiseq_type <- get_hiseq_type(x)
if(is.null(hiseq_type)) {
msg = paste0("\code{x} is not hiseq directory, ", x)
on.exit(message(msg), add = TRUE)
return(NULL)
} else if (! hiseq_type == "rnaseq_rx") {
msg = paste0("\code{x} is not hiseq_rx directory: ", x)
on.exit(message(msg), add = TRUE)
return(NULL)
}
pd        <- read_hiseq(x)
organism  <- pd$args$genome
deseq_dir <- pd$args$deseq_dir
fix_csv   <- file.path(deseq_dir, "transcripts_deseq2.fix.xls")
de_csv    <- file.path(deseq_dir, "transcripts_deseq2.csv")convert csv to xls

if(! file.exists(fix_csv)) {
message("Fix deseq csv: add mean, symbol, sig")
fix_df <- read_deseq_csv(de_csv, fc_cutoff = 2, pval_cutoff = 0.05,
readable = TRUE, organism = organism)
readr::write_delim(fix_df, fix_csv, delim = "\t", col_names = TRUE)
}output

if(file.exists(fix_csv)) {
fix_csv
} else {
stop(paste0("fix_csv file not found: ", x))
}
}}
\usage{
get_fix_csv(x)
}
\arguments{
\item{x}{character, path to the a.vs.b, rnaseq_rx directory}
}
\description{
#' @describeIn get_fix_csv Parsing the fix.xls file from DESeq2
#'
#'
#' @export
get_fix_csv <- function(x) {
if(is.character(x)) {
x <- x\link{1}
} else {
on.exit(message("'x' require character"), add = TRUE)
return(NULL)
}RNAseqRx directoryx/deseq/transcripts_deseq2.fix.xls

hiseq_type <- get_hiseq_type(x)
if(is.null(hiseq_type)) {
msg = paste0("\code{x} is not hiseq directory, ", x)
on.exit(message(msg), add = TRUE)
return(NULL)
} else if (! hiseq_type == "rnaseq_rx") {
msg = paste0("\code{x} is not hiseq_rx directory: ", x)
on.exit(message(msg), add = TRUE)
return(NULL)
}
pd        <- read_hiseq(x)
organism  <- pd$args$genome
deseq_dir <- pd$args$deseq_dir
fix_csv   <- file.path(deseq_dir, "transcripts_deseq2.fix.xls")
de_csv    <- file.path(deseq_dir, "transcripts_deseq2.csv")convert csv to xls

if(! file.exists(fix_csv)) {
message("Fix deseq csv: add mean, symbol, sig")
fix_df <- read_deseq_csv(de_csv, fc_cutoff = 2, pval_cutoff = 0.05,
readable = TRUE, organism = organism)
readr::write_delim(fix_df, fix_csv, delim = "\t", col_names = TRUE)
}output

if(file.exists(fix_csv)) {
fix_csv
} else {
stop(paste0("fix_csv file not found: ", x))
}
}
}
\section{Functions}{
\itemize{
\item \code{get_fix_csv()}: Parsing the fix.xls file from DESeq2

}}
