% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_utils.R
\name{get_kegg_keyType}
\alias{get_kegg_keyType}
\title{for clusterProfiler:
kegg,}
\usage{
get_kegg_keyType(organism)
}
\arguments{
\item{organism}{character name of the organism}
}
\description{
for KEGGREST
ncbi-geneid
}
\details{
for fruit fly: use flybasecg (OrgDb) => ncbi-geneid (kegg in clusterProfiler)
for human, mouse: use entrezid (OrgDb) => ncbi-geneid (kegg in clusterProfiler)
}
