% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{get_organism_name}
\alias{get_organism_name}
\title{get_organism_name}
\usage{
get_organism_name(x, group = "organism")
}
\arguments{
\item{x}{character the names used in other project}

\item{group}{character The name return, \link{"organism", "OrgDb", "TxDb"}}
}
\description{
get_organism_name
}
