% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{get_orgdb}
\alias{get_orgdb}
\title{get_orgdb}
\usage{
get_orgdb(x)
}
\arguments{
\item{x}{character The name of organism, or build name, eg: dm3, fruitfly
Support human, mouse and fruitfly

to-do: fetch from bioconductor
laod orgdb
GOSemSim::load_OrgDb()}
}
\description{
get_orgdb
}
