% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{get_orgdb_metadata}
\alias{get_orgdb_metadata}
\title{to-do: grepl() options, ignore.case}
\usage{
get_orgdb_metadata(x, options = "ORGANISM")
}
\arguments{
\item{x}{OrgDb}

\item{options}{character name of metadata, eg: ORGANISM}
}
\description{
to-do: grepl() options, ignore.case
}
\section{Functions}{
\itemize{
\item \code{get_orgdb_metadata()}: 

}}
