% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{get_orgdb_metadata}
\alias{get_orgdb_metadata}
\title{get_orgdb_info}
\usage{
get_orgdb_metadata(x, options = "ORGANISM")
}
\arguments{
\item{x}{OrgDb}

\item{options}{character name of metadata, eg: ORGANISM}
}
\description{
get_orgdb_info
}
