% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get_sig_gene}
\alias{get_sig_gene}
\title{get_sig_gene}
\usage{
get_sig_gene(data, fc_cutoff = 2, pval_cutoff = 0.05, type = "all")
}
\arguments{
\item{data}{data.frame, csv file}

\item{fc_cutoff}{numeric, cutoff for foldchange, default: 2}

\item{pval_cutoff}{numeric, cutoff for pvalue, default: 0.05}

\item{type}{string, all, both, up, down, not, default: all}
}
\description{
get_sig_gene
}
