% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{get_sig_name}
\alias{get_sig_name}
\title{get_sig_name}
\usage{
get_sig_name(x, ...)
}
\arguments{
\item{x}{data.frame, csv, xls output of \code{DESeq2::results(dds)}}

\item{fc}{numeric, cutoff for foldchange, default: 2}

\item{pvalue}{numeric, cutoff for pvalue, default: 0.05}

\item{p_adjust}{logical use p-adjust value instead}

\item{return_dataframe}{logical return a new data.frame, contain \code{sig} col}

\item{force}{logical force calculate sig, default: FALSE}

\item{.col_sig}{character name of the column for \code{sig}, default: "sig"}

\item{.col_log2fc}{character name of the log2fc column, default: "log2FoldChange"}

\item{.col_pvalue}{character name of the pvalue column, default: "pvalue";}

\item{.col_padj}{character name of the pvalue column, default: "padj";}

\item{.sig_up}{character assign name for up-regulated genes, default: "up"}

\item{.sig_down}{character assign name for down-regulated genes, default: "down"}

\item{.sig_not}{character assign name for not changed genes, default: "not"}
}
\value{
vector, sig names
}
\description{
get_sig_name
}
