% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peak.R
\name{get_txdb}
\alias{get_txdb}
\title{get_txdb}
\usage{
get_txdb(x)
}
\arguments{
\item{x}{character The name of organism, or build name, eg: dm3, fruitfly
Support human, mouse and fruitfly}
}
\description{
get_txdb
}
