% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_plot.R
\name{go_cnetplot}
\alias{go_cnetplot}
\title{create plots}
\usage{
go_cnetplot(x, ...)
}
\arguments{
\item{x}{object of enrichGO}

\item{..., }{support, fold_change, text_width,
fold_change
text_width
show_category}
}
\description{
create plots
}
