% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_plot.R
\name{go_emapplot_cluster}
\alias{go_emapplot_cluster}
\title{go_emapplot_cluster}
\usage{
go_emapplot_cluster(x, ...)
}
\arguments{
\item{x}{object of enrichGO

updated: 2020-12-15, pairwise_termsim(ego)
updated: 2021-09-01, deprecated, see treeplot()}
}
\description{
go_emapplot_cluster
}
