% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrich.R
\name{go_enrich}
\alias{go_enrich}
\title{Functions for GO analysis}
\usage{
go_enrich(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{character Genes for GO analysis}

\item{organism}{character Name of the organism, eg: dm3, fruitfly}

\item{...}{pass arguments: , orgdb, keytype, pval_cutoff, qval_cutoff,
readable, with default values:
orgdb = NULL
keytype = NULL
pval_cutoff = 0.9
qval_cutoff = 0.9
readable = TRUE}

\item{orgdb}{OrgDb from AnnotationDbi, overwrite \code{organism}}

\item{keytype}{character Name of the keytype for input, default: NULL}

\item{pval_cutoff}{float Cutoff for p-value, default: 0.9}

\item{qval_cutoff}{float Cutoff for q-value, default: 0.9}

\item{readable}{bool Args for groupGO function, convert to gene symbol

pval_cutoff, qval_cutoff, set to 0.9, in order to return
enrich results anyway, filter in downstream analysis}
}
\description{
organize GO output
generate tables, plots
}
\details{
go_group
\enumerate{
\item csv files, results
\item barplot,
\item wego plot,
}

go_enrich
\enumerate{
\item csv files, results
\item barplot
\item wego plot,
\item dotplot
\item netplot
\item ...
}

go_enrich
}
