% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_group.R
\name{go_group_plot}
\alias{go_group_plot}
\title{create plots}
\usage{
go_group_plot(x, ...)
}
\arguments{
\item{x}{object of enrichGO}

\item{...}{passing extra arguments
parent function: get_go_plots(),

text_width}
}
\description{
create plots
}
