% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_gsea.R
\name{go_gsea}
\alias{go_gsea}
\title{Functions for GO analysis}
\usage{
go_gsea(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{numeric}

\item{organism}{character Name of the organism, eg: dm3, fruitfly}

\item{...}{pass arguments: , orgdb, keytype, pval_cutoff, qval_cutoff,}

\item{fold_change}{numeric, with names}

\item{orgdb}{OrgDb from AnnotationDbi, overwrite \code{organism}}

\item{keytype}{character Name of the keytype for input, default: NULL}

\item{pval_cutoff}{float Cutoff for p-value, default: 0.9}
}
\description{
organize GO output
generate tables, plots
}
\details{
group_go
\enumerate{
\item csv files, results
\item barplot,
\item wego plot,
}

enrich_go
\enumerate{
\item csv files, results
\item barplot
\item wego plot,
\item dotplot
\item netplot
\item ...
}

go_gsea
}
