% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_plot.R
\name{go_heatplot}
\alias{go_heatplot}
\title{go_heatplot}
\usage{
go_heatplot(x, ...)
}
\arguments{
\item{x}{object of enrichGO}

\item{..., }{support, fold_change, text_width,
fold_change
text_width
show_category}
}
\description{
go_heatplot
}
