% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_plot.R
\name{go_plot
create plots}
\alias{go_plot
create plots}
\alias{go_barplot}
\title{Functions for generating plots for GO, KEGG analysis}
\usage{
go_barplot(x, ...)
}
\arguments{
\item{x}{list of objects of groupGO, enrichGO}

\item{..., }{support, fold_change, text_width,
fold_change
text_width
show_category}
}
\description{
Further plots are compatiable, with (gene_list, ...)
}
\details{
force "entrezid" in analysis pipeline
force "readable=T" for output
}
