% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{guess_keytype}
\alias{guess_keytype}
\title{guess_keytype}
\usage{
guess_keytype(x, organism = NULL, orgdb = NULL)
}
\arguments{
\item{x}{string, gene name}

\item{organism}{string, name of the organism, eg: Homo sapiens}
}
\description{
orgdb
}
