% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_deseq.R
\name{hiseq_deseq}
\alias{hiseq_deseq}
\title{Functions for DESeq2 down-stream analysis}
\usage{
hiseq_deseq(x, ...)
}
\arguments{
\item{x}{character path to the 'rnaseq_rx' directory}

\item{outdir}{character saving the results}

\item{strandness}{character could be "sens", "anti", default "sens"}

\item{fix_batch}{bool fix batch effect, default: TRUE}

\item{shrink}{logical \code{shrink} LFC by \link{"apeglm", "ashr", "normal"},
default: TRUE}

\item{transform}{logical transform dds by \code{vst()}, \code{rlog()}, default: TRUE}

\item{cpu}{integer, number of CPU to run in parallel, default: 4}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
DESeqResults, res (shrinked)
}
\description{
Reading files, directories, pipeline, ...
Processing data
Prepare for plot
Plotting
}
\details{
hiseq_deseq
}
