% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{hiseq_kraken2_report}
\alias{hiseq_kraken2_report}
\title{hiseq_kraken2_report}
\usage{
hiseq_kraken2_report(indir, outdir, preview = FALSE)
}
\arguments{
\item{indir}{character Path to the of kraken2 results}

\item{outdir}{character path to the html report}
}
\description{
hiseq_kraken2_report
}
