% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_p7.R
\name{hiseq_p7_report}
\alias{hiseq_p7_report}
\title{Functions for check P7 structure}
\usage{
hiseq_p7_report(input, output = NULL)
}
\description{
Check the structure of P7, TruSeq, Nextera
hiseq_p7_report
}
