% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{hiseq_prep_deseq4fc}
\alias{hiseq_prep_deseq4fc}
\title{Functions for DESeq2 analysis}
\usage{
hiseq_prep_deseq4fc(x, strandness = "sens", fix_batch = TRUE)
}
\arguments{
\item{x}{path to the directory of rnaseq_rx}

\item{strandness}{character could be "sens", "anti", default "sens"}

\item{fix_batch}{bool fix batch effect, default: TRUE}
}
\description{
Reading files, directories, pipeline, ...
Processing data
Prepare for plot
Plotting
}
\details{
hiseq_prep_deseq4fc
}
