% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{hiseq_prep_deseq4salmon}
\alias{hiseq_prep_deseq4salmon}
\title{hiseq_prep_deseq4salmon
Prepare data for DESeq analysis, from salmon results; rnaseq_rx}
\usage{
hiseq_prep_deseq4salmon(x, fix_batch = TRUE)
}
\arguments{
\item{x}{path to the directory of rnaseq_rx}

\item{fix_batch}{bool fix batch effect, default: TRUE}

\item{strandness}{character could be "sens", "anti", default "sens"}
}
\description{
hiseq_prep_deseq4salmon
Prepare data for DESeq analysis, from salmon results; rnaseq_rx
}
