% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_report.R
\name{hiseq_report
hiseq_report}
\alias{hiseq_report
hiseq_report}
\alias{hiseq_report}
\title{Functions for HiSeq report}
\usage{
hiseq_report(input, output = NULL, template_rmd = NULL)
}
\arguments{
\item{input}{directory to the sample}

\item{output}{directory to the html file}

\item{template}{the template, default from hiseqr}
}
\description{
Summary data in project_dir.
Processing data
Prepare for plot
Plotting
}
