% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_summary_modules.R
\name{hiseq_summary_modules}
\alias{hiseq_summary_modules}
\alias{read_hiseq_stat}
\title{Functions for HiSeq summary}
\usage{
read_hiseq_stat(x, keys = "align", add_tag = FALSE)
}
\arguments{
\item{x}{path to hiseq, single}

\item{keys}{character, quality control groups
options: c("trim", "align", "peak", "lendist", "frip", "report",
"enrich", "cor")}

\item{add_tag}{add SHA-256 value of dirnname(x), first-7 characters}
}
\description{
Summary data in project_dir.
Processing data
Prepare for plot
Plotting
}
