% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{import_featurecounts}
\alias{import_featurecounts}
\title{import_featurecounts Construct dds for DESeq2 analysis using matrix}
\usage{
import_featurecounts(x)
}
\arguments{
\item{x}{data.frame require "files", "smp_name", "condition" columns}
}
\value{
\code{DESeqDataSet}
}
\description{
using DESeqDataSetFromMatrix()
}
