% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{import_salmon}
\alias{import_salmon}
\title{import_samlmon
Construct dds for DESeq2 analysis using salmon output}
\usage{
import_salmon(x, tx2gene)
}
\arguments{
\item{x}{data.frame require "files", "names", "condition" columns}
}
\value{
\code{DESeqDataSet}
}
\description{
using DESeqDataSetFromMatrix()
}
