% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{is_go_result}
\alias{is_go_result}
\title{is_go_result}
\usage{
is_go_result(x, type = TRUE)
}
\arguments{
\item{x}{object}

\item{recursive}{boolean, whether loop over list}
}
\description{
is_go_result
}
