% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq.R
\name{is_hiseq_dir}
\alias{is_hiseq_dir}
\title{#' list_hiseq_dir:
#'
#' @param x path to the input directories
#' @param log boolen, whether print the log status, default: FALSE
#'
#' Fix: support x (multiple items)
#'
#' @export
list_hiseq_dir <- function(x, hiseq_type = "_r1"){
#--01.load data
pd <- read_hiseq(x)
if(! is(pd, "list")) {
# warning("Not a hiseq dir: x")
return(NULL)
}
#--02.get dirsgetattr, hasattr ?for alignment_rn/rx

if(is_hiseq_dir(x, "alignment")) {
x_dirs <- c(x, pd$args$rep_list)
x_dirs <- unique(x_dirs)
out <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type))
} else if(is_hiseq_dir(x, "_r1")) {
x_dirs <- x
out <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type))
} else if(is_hiseq_dir(x, "_rn")) {
x_dirs <- c(x, pd$args$rep_list) # r1 + rn
x_dirs <- unique(x_dirs)
out <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type))
} else if(is_hiseq_dir(x, "_rx")) {
# report: r1+rn+rx
if(is_hiseq_dir(x, "atac")) { # ATACseq
x_dirs <- list.dirs(x, recursive = FALSE)
x_dirs <- c(x, x_dirs)
x_dirs <- unique(x_dirs)
out    <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type = TRUE))
} else {
if(grepl("^chip|^cnt|^cnr", pd$hiseq_type, ignore.case = TRUE)) {
rn_dirs <- c(pd$args$ip_dir, pd$args$input_dir) # chip/cnr
} else if(is_hiseq_dir(x, "rnaseq_rx")) {
rn_dirs <- c(pd$args$mut_dir, pd$args$wt_dir) # RNAseq
} else {
rn_dirs <- c()
}
# for r1
r1 <- lapply(rn_dirs, list_hiseq_dir, hiseq_type="r1") \%>\% unlist()
# for rn
rn <- rn_dirs
# for rx
x_dirs <- unique(c(r1, rn, x))
# output
out <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type))
}
} else if(pd$is_hiseq_rp) {
out <- x
} else if(is_hiseq_dir(x, TRUE)) {
out <- x
} else {
out <- c()
}output

return(unique(out))
}
check x is hiseq or not}
\usage{
is_hiseq_dir(x, hiseq_type = TRUE)
}
\arguments{
\item{x}{string, path to dir}
}
\description{
#' list_hiseq_dir:
#'
#' @param x path to the input directories
#' @param log boolen, whether print the log status, default: FALSE
#'
#' Fix: support x (multiple items)
#'
#' @export
list_hiseq_dir <- function(x, hiseq_type = "_r1"){
#--01.load data
pd <- read_hiseq(x)
if(! is(pd, "list")) {
# warning("Not a hiseq dir: x")
return(NULL)
}
#--02.get dirsgetattr, hasattr ?for alignment_rn/rx

if(is_hiseq_dir(x, "alignment")) {
x_dirs <- c(x, pd$args$rep_list)
x_dirs <- unique(x_dirs)
out <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type))
} else if(is_hiseq_dir(x, "_r1")) {
x_dirs <- x
out <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type))
} else if(is_hiseq_dir(x, "_rn")) {
x_dirs <- c(x, pd$args$rep_list) # r1 + rn
x_dirs <- unique(x_dirs)
out <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type))
} else if(is_hiseq_dir(x, "_rx")) {
# report: r1+rn+rx
if(is_hiseq_dir(x, "atac")) { # ATACseq
x_dirs <- list.dirs(x, recursive = FALSE)
x_dirs <- c(x, x_dirs)
x_dirs <- unique(x_dirs)
out    <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type = TRUE))
} else {
if(grepl("^chip|^cnt|^cnr", pd$hiseq_type, ignore.case = TRUE)) {
rn_dirs <- c(pd$args$ip_dir, pd$args$input_dir) # chip/cnr
} else if(is_hiseq_dir(x, "rnaseq_rx")) {
rn_dirs <- c(pd$args$mut_dir, pd$args$wt_dir) # RNAseq
} else {
rn_dirs <- c()
}
# for r1
r1 <- lapply(rn_dirs, list_hiseq_dir, hiseq_type="r1") \%>\% unlist()
# for rn
rn <- rn_dirs
# for rx
x_dirs <- unique(c(r1, rn, x))
# output
out <- purrr::keep(x_dirs, is_hiseq_dir(x_dirs, hiseq_type))
}
} else if(pd$is_hiseq_rp) {
out <- x
} else if(is_hiseq_dir(x, TRUE)) {
out <- x
} else {
out <- c()
}output

return(unique(out))
}
check x is hiseq or not
}
