% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{is_valid_index}
\alias{is_valid_index}
\title{check index id}
\usage{
is_valid_index(x, hiseq_type = "p7", skip_null = TRUE)
}
\arguments{
\item{x}{character}

\item{hiseq_type}{character, p7, barcode}
}
\description{
check index id
}
