% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{is_valid_kegg_input}
\alias{is_valid_kegg_input}
\title{check_go_input}
\usage{
is_valid_kegg_input(...)
}
\description{
Required:
\itemize{
\item gene_list, chr, characters (group, enrich)
\item outdir, chr, character
\item organism, chr, character
\item orgdb, OrgDb, AnnotationDbi, one of organism, orgdb required
}

optional:
\itemize{
\item for_gsea, logical, gene_list, num, sorted, named (GSEA)
}
}
