% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{is_valid_keys}
\alias{is_valid_keys}
\title{is_valid_keys}
\usage{
is_valid_keys(orgdb, keys, keytype, fks = NULL, return_pct = FALSE)
}
\arguments{
\item{orgdb}{OrgDb}

\item{keys}{character gene names}

\item{keytype}{character keytype}

\item{fks}{vector}
}
\description{
fks is an alternate vector of keys to consult for validity.
Normally this will be NULL and the test function should consult
keys for the supplied keytype
}
