% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R, R/kegg_gsea.R
\name{kegg}
\alias{kegg}
\alias{kegg_gsea}
\title{Functions for KEGG analysis}
\usage{
kegg(gene_list, organism, ...)

kegg_gsea(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{decreasing sorted numeric vector}

\item{organism}{kegg_code, eg: dme}

\item{...}{extra argument

fold_change, pval_cutoff, qval_cutoff}

\item{outdir}{string}
}
\description{
overrepresent analysis
GSEA analysis

overrepresent analysis
GSEA analysis
}
\details{
enrich kegg
\enumerate{
\item csv files, results
\item barplot
\item wego plot,
\item dotplot
\item netplot
\item ...
}

gsea kegg
\enumerate{
\item csv files, results
\item barplot
\item dotplot
\item netplot
\item ...
}

gsea kegg
\enumerate{
\item csv files, results
\item barplot
\item dotplot
\item netplot
\item ...
}
}
