% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_gsea.R
\name{kegg_gsea}
\alias{kegg_gsea}
\title{Functions for KEGG analysis}
\usage{
kegg_gsea(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{decreasing sorted numeric vector}

\item{organism}{kegg_code, eg: dme}
}
\description{
overrepresent analysis
GSEA analysis
}
\details{
gsea kegg
\enumerate{
\item csv files, results
\item barplot
\item dotplot
\item netplot
\item ...
}

kegg_gsea
}
