% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_utils.R
\name{kegg_utils}
\alias{kegg_utils}
\alias{to_kegg_gene_id}
\title{Functions for KEGG analysis}
\usage{
to_kegg_gene_id(
  gene_list,
  organism,
  keytype = NULL,
  simplified = TRUE,
  rm_na = TRUE,
  return_table = FALSE
)
}
\arguments{
\item{gene_list}{character geneid}

\item{organism}{character Name of the organism, support abbr for fruitfly,
human, mouse, eg: dm3, dm6, hg38, mouse, mm10, human,}

\item{keytype}{character or NULL The keytype of gene_list, if NULL,
will guess the keytype from keytypes(OrgDb)}
}
\description{
Convert between KEGG id and OrgDb id

for most organism:
{kegg_code}:{OrgDb_entrezid}

for fruitfly:
{kegg_code}:{Dmel_}{OrgDb_FLYBASECG}
keytype -> FLYBASECG -> dme:Dmel_{CG10000}

for prokaryotes
do not contains: entrezid, use uniprot instead
{up}:{uniprot_id}

KEGG supported gene type:
ncbi-geneid, ncbi-proteinid, uniprot, flybase (for dme)
}
\details{
Usage: KEGGREST functions:
keggInfo(), keggList(), keggFind(), keggGet(), keggConv, and keggLink()
\enumerate{
\item list all supprted organism
keggList("organism")
\item list genes, for one organism
keggList("dme")
\item convert ids (idtype)
keggConv("ncbi-proteinid", c("hsa:10458", "ece:Z5100"))
head(keggConv("eco", "ncbi-geneid")) # all entries
}
}
