% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{list_align_arguments_file}
\alias{list_align_arguments_file}
\title{list_arguments_file}
\usage{
list_align_arguments_file(x)
}
\arguments{
\item{x}{path to the directory.
expect x/config/{feature}/config/arguments.pickle,
x/{feature}/config/arguments.pickle}
}
\description{
list_arguments_file
}
