% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq.R
\name{list_arguments_file}
\alias{list_arguments_file}
\title{list_arguments_file Search config files, .pickle, .toml
list_arguments_file}
\usage{
list_arguments_file(x)
}
\arguments{
\item{x}{path to the directory.
expect
x/config/arguments.pickle
x/{feature}/config/arguments.pickle
x/config/{feature}/config/arguments.pickle,

filename could be config.pickle}
}
\description{
list_arguments_file Search config files, .pickle, .toml
list_arguments_file
}
