% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacseq.R
\name{list_atac_multiple_dirs}
\alias{list_atac_multiple_dirs}
\title{atacseq_merge_dirs}
\usage{
list_atac_multiple_dirs(x)
}
\arguments{
\item{x}{path to the input directories}
}
\description{
atacseq_merge_dirs
}
