% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacseq.R
\name{list_atac_single_dirs}
\alias{list_atac_single_dirs}
\title{list_atac_single_dirs}
\usage{
list_atac_single_dirs(x)
}
\arguments{
\item{x}{path to the input directories}

\item{log}{boolen, whether print the log status, default: FALSE}
}
\description{
list_atac_single_dirs
}
