% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{list_chipseq_merge_dirs}
\alias{list_chipseq_merge_dirs}
\title{chipseq_merge_dirs}
\usage{
list_chipseq_merge_dirs(x)
}
\arguments{
\item{x}{path to the input directories}
}
\description{
chipseq_merge_dirs
}
