% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq.R
\name{list_hiseq_dir}
\alias{list_hiseq_dir}
\title{Functions for HiSeq pipeline}
\usage{
list_hiseq_dir(x, hiseq_type = "auto")
}
\arguments{
\item{x}{path to the input directories}

\item{log}{boolen, whether print the log status, default: FALSE

Fix: support x (multiple items)}
}
\description{
Reading files, directories, pipeline, ...
Processing data
Prepare for plot
Plotting
}
\details{
list_hiseq_dir:
}
