% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq.R
\name{list_hiseq_file}
\alias{list_hiseq_file}
\title{list_hiseq_file}
\usage{
list_hiseq_file(x, keys = "bam", hiseq_type = "auto")
}
\arguments{
\item{x}{path to the input directories}

\item{keys}{character, attributes/names in hiseq_dir, default: \link{"bam"}}
}
\description{
list_hiseq_file
}
