% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{load_hiseq_index}
\alias{load_hiseq_index}
\title{load index}
\usage{
load_hiseq_index(hiseq_type = TRUE)
}
\arguments{
\item{hiseq_type}{character, \link{truseq, nextera, barcode}}
}
\description{
load index
}
