% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{load_hiseq_sheet_dir}
\alias{load_hiseq_sheet_dir}
\title{Example:
load hiseq sample sheet files}
\usage{
load_hiseq_sheet_dir(x)
}
\arguments{
\item{x}{string}
}
\value{
data.frame
}
\description{
Example:
load hiseq sample sheet files
}
