% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ma_plot}
\alias{ma_plot}
\title{ma_plot Create MA plot for deseq output}
\usage{
ma_plot(
  data,
  x = "baseMean",
  y = "log2FoldChange",
  labels = NULL,
  title = "MA plot"
)
}
\arguments{
\item{data}{A data frame for the full version of data, required}

\item{x}{character Column name show on x-axis, eg: baseMean}

\item{y}{character Column name chow on y-axis, eg: log2FoldChange}

\item{labels}{character Labels to show in plot}

\item{title}{character Title, default: "MA plot"}
}
\description{
ma_plot Create MA plot for deseq output
}
