% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{make_publish_plots}
\alias{make_publish_plots}
\title{deprecated: see deseq_utils.R}
\usage{
make_publish_plots(x, outdir = NULL, to_pdf = TRUE)
}
\arguments{
\item{x}{character rnaseq_rx directory, a.vs.b}

\item{outdir}{character Path to the directory, saving the plots}

\item{save2pdf}{bool Save the plots in PDF file,}
}
\description{
deprecated: see deseq_utils.R
}
\section{Functions}{
\itemize{
\item \code{make_publish_plots()}: Create plots with pubilsh quality

}}
\section{sample names}{
rep_name <- data \%>\%
dplyr::select(Gene:baseMean) \%>\%
dplyr::select(-Gene, -baseMean) \%>\%
names()
}

\section{groups}{

}

