% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffbind.R
\name{np2bed4}
\alias{np2bed4}
\title{Convert narrowPeak to bed4}
\usage{
np2bed4(np, outdir = NULL)
}
\arguments{
\item{np}{narrowPeak file}

\item{outdir}{path to save the bed file, if NULL, save to the same directory
of np}
}
\description{
np2bed4
}
\details{
Convert narrowPeak to BED4

save column-7: signalValue
}
