% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_to_list}
\alias{path_to_list}
\title{path_to_list}
\usage{
path_to_list(path, recursive = FALSE)
}
\arguments{
\item{path}{character Path to the file or directory}

\item{recursive}{bool Constructure the nested list recursive}
}
\description{
path_to_list
}
