% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{plot_fastqc}
\alias{plot_fastqc}
\title{Functions for fastqc}
\usage{
plot_fastqc(r1, r2 = NULL, module = "Per base sequence content", ...)
}
\arguments{
\item{r1}{List output of \code{qc_read()} of read1}

\item{r2}{List output of \code{qc_read()} of read2}

\item{module}{Character name of the modules}
}
\description{
plot_fastqc
}
