% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{plot_heatmap_deseq_pair}
\alias{plot_heatmap_deseq_pair}
\title{to-do: hclust rows !!!!
RNAseq cmp,
number of sig genes}
\usage{
plot_heatmap_deseq_pair(x = NULL, y = NULL, .pd_rds = NULL, group = "all", ...)
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{.pd_rds}{string path to the file, saving read_deseq_pair() output}

\item{feature}{string default: gene}
}
\description{
to-do: hclust rows !!!!
RNAseq cmp,
number of sig genes
}
