% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.hiseq_plots2.R, R/hiseq_plots.R
\name{plot_hiseq_align}
\alias{plot_hiseq_align}
\title{plot_rnaseq_align Create bar_plot for align_stat}
\usage{
plot_hiseq_align(
  x,
  mode = 1,
  columns = NULL,
  title = "Alignment",
  add_label = FALSE,
  position = "fill",
  fish = "Trimma_lantana",
  ...
)

plot_hiseq_align(
  x,
  mode = 1,
  columns = NULL,
  title = "Alignment",
  add_label = FALSE,
  position = "fill",
  fish = "Trimma_lantana",
  ...
)
}
\arguments{
\item{mode}{integer map=1, unique/multiple=2, default: 1}

\item{fish}{string Name of the fish, use \code{fishualize::fish_palettes()} list
all availabel fish names}

\item{data}{data.frame From get_rnaseq_align_stat()}
}
\description{
Output from get_rnaseq_align_stat()
including columns:
fqname, map, unique, multiple

Output from get_rnaseq_align_stat()
including columns:
fqname, map, unique, multiple
}
