% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.hiseq_plots2.R, R/hiseq_plots.R
\name{plot_hiseq_bam_cor}
\alias{plot_hiseq_bam_cor}
\title{plot_hiseq_bam_cor}
\usage{
plot_hiseq_bam_cor(x)

plot_hiseq_bam_cor(x)
}
\arguments{
\item{x}{path to the directory}
}
\description{
generate bam cor plots for hiseq dirs

generate bam cor plots for hiseq dirs
}
