% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_plots.R
\name{plot_hiseq_frip}
\alias{plot_hiseq_frip}
\title{plot_hiseq_frip Create bar_plot for FRiP}
\usage{
plot_hiseq_frip(x, hiseq_type = TRUE, title = "FRiP")
}
\arguments{
\item{data}{data.frame From get_rnaseq_align_stat()}

\item{mode}{integer map=1, unique/multiple=2, default: 1}

\item{fish}{string Name of the fish, use \code{fishualize::fish_palettes()} list
all availabel fish names}
}
\description{
Output from read_hiseq_frip_stat()
including columns:
filename, total, map, pct, label (pct=FRiP)
}
