% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.hiseq_plots2.R, R/hiseq_plots.R
\name{plot_hiseq_lendist}
\alias{plot_hiseq_lendist}
\title{plot_hiseq_peak Create bar_plot for align_stat}
\usage{
plot_hiseq_lendist(
  x,
  hiseq_type = TRUE,
  title = "Distribution of fragment size",
  fish = "Trimma_lantana"
)

plot_hiseq_lendist(
  x,
  hiseq_type = TRUE,
  title = "Distribution of fragment size",
  fish = "Trimma_lantana"
)
}
\arguments{
\item{fish}{string Name of the fish, use \code{fishualize::fish_palettes()} list
all availabel fish names}

\item{data}{data.frame From get_rnaseq_align_stat()}

\item{mode}{integer map=1, unique/multiple=2, default: 1}
}
\description{
Output from get_rnaseq_align_stat()
including columns:
fqname, map, unique, multiple

Output from get_rnaseq_align_stat()
including columns:
fqname, map, unique, multiple
}
