% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.hiseq_plots2.R, R/hiseq_plots.R
\name{plot_hiseq_trim}
\alias{plot_hiseq_trim}
\title{plot_hiseq_trim Create bar_plot for trim stat}
\usage{
plot_hiseq_trim(
  x,
  fish = "Trimma_lantana",
  direction = "horizontal",
  position = "fill"
)

plot_hiseq_trim(
  x,
  fish = "Trimma_lantana",
  direction = "horizontal",
  position = "fill"
)
}
\arguments{
\item{x}{data.frame From get_trim_stat}

\item{fish}{string Name of the fish, use \code{fishualize::fish_palettes()} list}

\item{direction}{string \link{"horizontal", "vertical"}}

\item{position}{string \link{"stack", "fill"}
all available fish names}
}
\description{
output from get_rnaseq_trim_stat(),
including columns:
id, raw, clean, clean_pct, short_pct

output from get_rnaseq_trim_stat(),
including columns:
id, raw, clean, clean_pct, short_pct
}
