% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{plot_kraken2_report}
\alias{plot_kraken2_report}
\title{plot_kraken2_report
Generate heatmap for kraken2 report}
\usage{
plot_kraken2_report(x, tax_level = "G", topN = 5, switch_axis = FALSE)
}
\arguments{
\item{x}{character, path to the kraken2 output

A rank code, indicating (U)nclassified, (R)oot, (D)omain, (K)ingdom,
(P)hylum, (C)lass, (O)rder, (F)amily, (G)enus, or (S)pecies. Taxa that are
not at any of these 10 ranks have a rank code that is formed by using the
rank code of the closest ancestor rank with a number indicating the distance
from that rank. E.g., "G2" is a rank code indicating a taxon is between}
}
\description{
plot_kraken2_report
Generate heatmap for kraken2 report
}
