% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{plot_overlap_deseq_pair}
\alias{plot_overlap_deseq_pair}
\title{plot_overlap_deseq_pair}
\usage{
plot_overlap_deseq_pair(x = NULL, y = NULL, .pd_rds = NULL, group = "all", ...)
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{.pd_rds}{string path to the file, saving read_deseq_pair() output}

\item{...}{parameters for deseq_qc_res()}
}
\description{
plot_overlap_deseq_pair
}
