% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peak.R
\name{plot_peak_annotation}
\alias{plot_peak_annotation}
\title{make plot}
\usage{
plot_peak_annotation(
  x,
  genome = NULL,
  tss_region = 3000,
  plot_type = "bar",
  fish = "Trimma_lantana",
  ...
)
}
\arguments{
\item{x}{csAnno object or list of csAnno}

\item{plot_type}{string, bar, tss, venn, upset}
}
\description{
make plot
}
