% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peak.R
\name{plot_peak_profile}
\alias{plot_peak_profile}
\title{plot_peak_profile}
\usage{
plot_peak_profile(
  x,
  genome = NULL,
  flanking = 2000,
  bed = NULL,
  on_summit = FALSE,
  facet = "none",
  conf = NA,
  fish = "Trimma_lantana",
  ...
)
}
\description{
plot_peak_profile
}
