% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.hiseq_plots2.R, R/hiseq_plots.R
\name{plot_rnaseq_featureCounts}
\alias{plot_rnaseq_featureCounts}
\title{plot_rnaseq_featureCounts}
\usage{
plot_rnaseq_featureCounts(x, fish = "Trimma_lantana")

plot_rnaseq_featureCounts(x, fish = "Trimma_lantana")
}
\arguments{
\item{x}{character path to RNAseq r1, rn}
}
\description{
plot_rnaseq_featureCounts

plot_rnaseq_featureCounts
}
