% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_plots.R
\name{plot_rnaseq_sig_count}
\alias{plot_rnaseq_sig_count}
\title{plot_rnaseq_sig_count for significant genes}
\usage{
plot_rnaseq_sig_count(x, fish = "Trimma_lantana")
}
\arguments{
\item{x}{character path to the rnaseq_rx}
}
\description{
plot_rnaseq_sig_count for significant genes
}
