% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.hiseq_plots2.R
\name{plot_rnaseq_sig_count2}
\alias{plot_rnaseq_sig_count2}
\title{plot_rnaseq_sig_count for significant genes}
\usage{
plot_rnaseq_sig_count2(
  x,
  shrink_method = "standard",
  skip_not = TRUE,
  return_data = FALSE,
  fish = "Trimma_lantana"
)
}
\arguments{
\item{x}{character path to the rnaseq_rx}
}
\description{
plot_rnaseq_sig_count for significant genes
}
