% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{plot_seq_quality_scores}
\alias{plot_seq_quality_scores}
\title{plot_seq_quality_scores}
\usage{
plot_seq_quality_scores(x, ...)
}
\arguments{
\item{x}{List output of \code{qc_read(x, module = TRUE)}}
}
\description{
plot_seq_quality_scores
}
