% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{plot_tile_seq_quality}
\alias{plot_tile_seq_quality}
\title{plot_tile_seq_quality}
\usage{
plot_tile_seq_quality(x, ...)
}
\arguments{
\item{x}{List output of \code{qc_read(x, module = TRUE)}}
}
\description{
plot_tile_seq_quality
}
