% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{prep_go}
\alias{prep_go}
\title{prep_go}
\usage{
prep_go(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{gene names}

\item{organism}{string, eg: Homo sapiens}

\item{...}{extra argument}

\item{fold_change}{numeric, fold change}

\item{keytype}{character}
}
\description{
prep_go
}
