% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_gsea.R
\name{prep_go_gsea}
\alias{prep_go_gsea}
\title{prep_go_gsea}
\usage{
prep_go_gsea(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{character, gene name}

\item{organism}{character, name of the genome}

\item{fold_change}{numeric, named}
}
\description{
prep_go_gsea
}
