% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{prep_kegg}
\alias{prep_kegg}
\title{prep_kegg}
\usage{
prep_kegg(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{gene names}

\item{organism}{string, eg: Homo sapiens}

\item{...}{extra argument}

\item{fold_change}{numeric, fold change}

\item{keytype}{character}
}
\description{
prep_kegg
}
