% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_gsea.R
\name{prep_kegg_gsea}
\alias{prep_kegg_gsea}
\title{prep_kegg_gsea}
\usage{
prep_kegg_gsea(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{gene name}

\item{fold_change}{numeric}
}
\description{
prep_kegg_gsea
}
