% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{read_deseq_csv}
\alias{read_deseq_csv}
\title{read_deseq_csv}
\usage{
read_deseq_csv(
  x,
  fc_cutoff = 2,
  pval_cutoff = 0.05,
  readable = TRUE,
  organism = NULL
)
}
\arguments{
\item{x}{string Path to the .csv file}

\item{fc_cutoff}{numeric foldChange cutoff, default: 2}

\item{pval_cutoff}{numeric pvalue cutoff, default: 0.05}
}
\description{
read_deseq_csv
}
