% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{read_deseq_pair}
\alias{read_deseq_pair}
\title{RNAseq cmp, stat
number of sig genes}
\usage{
read_deseq_pair(x, y, ...)
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{...}{extra parameters for deseq_qc_res()}
}
\description{
RNAseq cmp, stat
number of sig genes
}
