% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_fc}
\alias{read_fc}
\title{read_fc}
\usage{
read_fc(x, digits = 0, get_gene_length = FALSE)
}
\arguments{
\item{x}{path to count.txt file, featureCounts output}

\item{digits}{integers, Number of digits keep, default: 0}

\item{get_gene_length}{bool, save gene_lengthin column 5}
}
\description{
read_fc
}
