% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_fc_summary}
\alias{read_fc_summary}
\title{read_fc_summary}
\usage{
read_fc_summary(x, fix_name = TRUE)
}
\arguments{
\item{x}{path to the count.txt.summary file}

\item{fix_name}{bool extract file name,}
}
\description{
read_fc_summary
}
