% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq.R
\name{read_hiseq}
\alias{read_hiseq}
\title{#' check x is hiseq or not
#' @param x string, path to dir
#'
#' @export
#'
#' @example
#'
#' > is_hiseq_dir(x, "r1")
#' > is_hsieq_dir(x, "rnaseq")
#' > is_hiseq_dir(x, "atacseq_rx")
#'
#'
is_hiseq_dir <- function(x, hiseq_type = TRUE) {
if(inherits(x, "character")) {
sapply(x, function(i) {
pd     <- read_hiseq(i)
i_type <- ifelse(is(pd, "list"), pd$hiseq_type, "")
i_is_hiseq <- ifelse(is(pd, "list"), pd$is_hiseq, FALSE)
i_is_hiseq <- i_is_hiseq && (! endsWith(i, "config")) # exclude "config/"
if(isTRUE(hiseq_type)) {
i_is_hiseq
} else if(is(hiseq_type, "character")) {
endsWith(i_type, hiseq_type) | startsWith(i_type, hiseq_type)
}
})
} else {
FALSE
}
}
read HiSeq directory}
\usage{
read_hiseq(x)
}
\arguments{
\item{x}{string path

Fix: support multiple element}
}
\description{
input dir
output files: pickle, toml, json
}
