% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_summary_modules.R
\name{read_hiseq_align_stat}
\alias{read_hiseq_align_stat}
\title{read_hiseq_align_stat}
\usage{
read_hiseq_align_stat(x)
}
\arguments{
\item{x}{path to hiseq, single}
}
\description{
read_hiseq_align_stat
}
