% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_summary_modules.R
\name{read_hiseq_lendist_json}
\alias{read_hiseq_lendist_json}
\title{read_hiseq_lendist_json}
\usage{
read_hiseq_lendist_json(x)
}
\arguments{
\item{x}{path to the csv/json file}
}
\description{
read_hiseq_lendist_json
}
