% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{read_hiseq_sheet}
\alias{read_hiseq_sheet}
\title{read sample info from Excel.xlsx}
\usage{
read_hiseq_sheet(x, n_max = 10000, fix = TRUE, sampleid_start = 1)
}
\arguments{
\item{x}{file xlsx}
}
\description{
read sample info from Excel.xlsx
}
