% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{read_hiseq_trim}
\alias{read_hiseq_trim}
\title{functions for hiseq trim}
\usage{
read_hiseq_trim(x)
}
\arguments{
\item{x}{character or data.frame}
}
\description{
read data from hiseq.trim output
}
\details{
read_hiseq_trim
}
