% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_summary_modules.R
\name{read_hiseq_trim_stat}
\alias{read_hiseq_trim_stat}
\title{read_hiseq_trim_stat}
\usage{
read_hiseq_trim_stat(x)
}
\arguments{
\item{x}{path to hiseq, single}
}
\description{
read_hiseq_trim_stat
}
