% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{read_kraken2_report}
\alias{read_kraken2_report}
\title{read_kraken2_report}
\usage{
read_kraken2_report(x, tax_level = "G", topN = 5)
}
\arguments{
\item{x}{character path to the directory of kraken2 results}

\item{tax_level}{character the level of tax, default; \link{G}}

\item{topN}{int the number of taxon to show

choose topN taxon across all samples
assign the rest as "other"

A rank code, indicating (U)nclassified, (R)oot, (D)omain, (K)ingdom,
(P)hylum, (C)lass, (O)rder, (F)amily, (G)enus, or (S)pecies. Taxa that are
not at any of these 10 ranks have a rank code that is formed by using the
rank code of the closest ancestor rank with a number indicating the distance
from that rank. E.g., "G2" is a rank code indicating a taxon is between
genus and species and the grandparent taxon is at the genus rank}
}
\description{
read_kraken2_report
}
