% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peak.R
\name{read_peak_annotation}
\alias{read_peak_annotation}
\title{read_peak_annotation}
\usage{
read_peak_annotation(x, genome = NULL, tss_region = 3000, ...)
}
\arguments{
\item{x}{peak file or hiseq_dir path}

\item{genome}{string, optional if x is hiseq_dir}

\item{tss_region}{numeric, 3000}

\item{...}{more arguments}
}
\description{
read_peak_annotation
}
