% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peak.R
\name{read_peak_profile}
\alias{read_peak_profile}
\title{read_peak_profile}
\usage{
read_peak_profile(
  x,
  genome = NULL,
  flanking = 2000,
  bed = NULL,
  on_summit = FALSE
)
}
\arguments{
\item{x}{peak file or hiseq_dir path}

\item{genome}{string, optional if x is hiseq_dir}

\item{tss_region}{numeric, 3000}
}
\description{
read_peak_profile
}
