% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_summary_modules.R
\name{read_rnaseq_deseq}
\alias{read_rnaseq_deseq}
\title{read_rnaseq_deseq}
\usage{
read_rnaseq_deseq(x, sig_list = TRUE)
}
\arguments{
\item{x}{character path to the rnaseq_rx}
}
\description{
read_rnaseq_deseq
}
