% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.hiseq_plots2.R
\name{read_rnaseq_deseq2}
\alias{read_rnaseq_deseq2}
\title{read_rnaseq_deseq2
for upgraded version deseq, including shrink results}
\usage{
read_rnaseq_deseq2(x, shrink_method = "standard", sig_type = "all")
}
\arguments{
\item{x}{character path to the rnaseq_rx}
}
\description{
read_rnaseq_deseq2
for upgraded version deseq, including shrink results
}
