% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq}
\alias{rnaseq}
\alias{rnaseq_hub}
\alias{deseq2_main}
\title{Functions for RNAseq analysis}
\usage{
rnaseq_hub(x, ...)

deseq2_main(
  dds,
  outdir,
  fc_cutoff = 2,
  pval_cutoff = 0.05,
  organism = NULL,
  readable = TRUE
)
}
\arguments{
\item{x}{path to the RNAseqRx directory, a.vs.b}

\item{...}{extra arguments for demseq2_main function#'

deseq2_main() extra arguments:
\itemize{
\item fc_cutoff    : 2
\item pval_cutoff  : 0.05
\item readable     : TRUE
}}

\item{dds}{count data in matrix}

\item{outdir}{Directory to save the results}

\item{fc_cutoff}{float Cutoff for foldchange, default: 2}

\item{pval_cutoff}{float Cutoff for pvalue, default: 0.05}

\item{organism}{character, the name of the organism, eg: dm6, hg38, human,}

\item{readable}{bool, Add symbol and entrezid to table}
}
\description{
Apply DESeq2 for RNAseq DE analysis
require >=2 replicates for each sample
}
\details{
input: count.txt, from featureCounts output
output: DESeq2 output

helper functions:
\enumerate{
\item read_fc
}
}
\section{Functions}{
\itemize{
\item \code{deseq2_main()}: The main function do the analysis

}}
