% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq_enrich_hub}
\alias{rnaseq_enrich_hub}
\title{rnaseq_enrich}
\usage{
rnaseq_enrich_hub(x, ...)
}
\arguments{
\item{x}{path to the RNAseqRx directory: a.vs.b}

\item{...}{extra arguments for demseq2_main function

extra arguments:
\itemize{
\item overwrite    : FALSE
\item text_width   : 40
\item pval_cutoff  : 0.9,
\item qval_cutoff  : 0.9,
\item level        : 2
\item readable     : bool
\item layout       : nicely, kk
\item show_category: 12
}}
}
\description{
rnaseq_enrich
}
