% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq_hub}
\alias{rnaseq_hub}
\title{Functions for RNAseq analysis}
\usage{
rnaseq_hub(x, ...)
}
\arguments{
\item{x}{path to the RNAseqRx directory, a.vs.b}

\item{...}{extra arguments for demseq2_main function#'

deseq2_main() extra arguments:
\itemize{
\item fc_cutoff    : 2
\item pval_cutoff  : 0.05
\item readable     : TRUE
}}

\item{organism}{character, the name of the organism, eg: dm6, hg38, human,}

\item{readable}{bool, Add symbol and entrezid to table
organism, fc_cutoff, pval_cutoff, readable,}
}
\description{
Apply DESeq2 for RNAseq DE analysis
require >=2 replicates for each sample
}
\details{
input: count.txt, from featureCounts output
output: DESeq2 output

helper functions:
\enumerate{
\item read_fc
}

rnaseq_hub
}
