% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq_report}
\alias{rnaseq_report}
\title{rnaseq_report}
\usage{
rnaseq_report(input, output, template_rmd = NULL)
}
\arguments{
\item{input}{directory to the sample}

\item{output}{directory to the html file}

\item{template}{the template, default from hiseqr}
}
\description{
rnaseq_report
}
