% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rnaseq_sig_stat_plot}
\alias{rnaseq_sig_stat_plot}
\title{rnaseq_sig_stat_plot Create bar_plot for sig count}
\usage{
rnaseq_sig_stat_plot(data, fish = "Trimma_lantana")
}
\arguments{
\item{data}{data.frame From deseq_dir, fix.xls}
}
\description{
rnaseq_sig_stat_plot Create bar_plot for sig count
}
