% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq.R
\name{run_deseq_res}
\alias{run_deseq_res}
\title{run_deseq_res}
\usage{
run_deseq_res(dds, ...)
}
\arguments{
\item{dds}{DESeqDataSet}

\item{outdir}{character saving deseq_res to file: deseq_res.rds}

\item{shrink}{logical \code{shrink} LFC by \link{"apeglm", "ashr", "normal"},
default: TRUE}

\item{transform}{logical transform dds by \code{vst()}, \code{rlog()}, default: TRUE}

\item{cpu}{integer, number of CPU to run in parallel, default: 2}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
list(dds=, dds_trans = list(), res=, res_lfc=list())
}
\description{
run_deseq_res
}
