% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon.R
\name{salmon}
\alias{salmon}
\alias{rnaseq_salmon_hub}
\title{Functions for salmon output}
\usage{
rnaseq_salmon_hub(x, ...)
}
\arguments{
\item{x}{path to the RNAseqRx directory, a.vs.b}

\item{...}{extra arguments for demseq2_main function#'

deseq2_main() extra arguments:
\itemize{
\item fc_cutoff    : 2
\item pval_cutoff  : 0.05
\item readable     : TRUE
}}

\item{organism}{character, the name of the organism, eg: dm6, hg38, human,}

\item{readable}{bool, Add symbol and entrezid to table
organism, fc_cutoff, pval_cutoff, readable,}
}
\description{
Apply DESeq2 for RNAseq DE analysis
}
