% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scatter_plot2}
\alias{scatter_plot2}
\title{scatter_plot2 Create scatter plot, version2}
\usage{
scatter_plot2(data, x, y, ...)
}
\arguments{
\item{data}{A data frame for the full version of data, required}

\item{x}{character Column name show on x-axis}

\item{y}{character Column name chow on y-axis}

\item{labels}{character Labels to show in plot}

\item{add_label_point}{bool Add points for labels}

\item{show_abline}{bool Add ablines}

\item{trans}{character Function to convert num log10 or log2}

\item{xmin}{float Min on x-axis}

\item{xmax}{float Max x on x-axis}

\item{ymin}{float Min on y-axis}

\item{ymax}{float Max y on y-axis}

\item{title}{character Title, default: "Scatter plot"}
}
\description{
scatter_plot2 Create scatter plot, version2
}
