% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{set_readable}
\alias{set_readable}
\title{set_readable}
\usage{
set_readable(x, ...)
}
\arguments{
\item{x}{data.frame contains \code{gene_id}
alternative,\code{Gene}, \code{id}, or \code{rown.names()}}

\item{genome}{character name of the genome,
eg: dm6, hg38, mm10}
}
\value{
data.frame
}
\description{
set_readable
}
