% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_DT}
\alias{to_DT}
\title{to_DT}
\usage{
to_DT(df, mode = 1, pageLength = 10)
}
\arguments{
\item{df}{data.frame}

\item{mode}{integer default 1}

\item{page_length}{integer default 10

see https://rstudio.github.io/DT/ for details}
}
\description{
to_DT
}
