% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_utils.R
\name{to_kegg_gene_id}
\alias{to_kegg_gene_id}
\title{Functions for KEGG analysis}
\usage{
to_kegg_gene_id(
  gene_list,
  organism,
  keytype = NULL,
  simplified = TRUE,
  rm_na = TRUE,
  return_table = FALSE
)
}
\arguments{
\item{gene_list}{character geneid}

\item{organism}{character Name of the organism, support abbr for fruitfly,
human, mouse, eg: dm3, dm6, hg38, mouse, mm10, human,}

\item{keytype}{character or NULL The keytype of gene_list, if NULL,
will guess the keytype from keytypes(OrgDb)}
}
\description{
Usage: KEGGREST functions:
keggInfo(), keggList(), keggFind(), keggGet(), keggConv, and keggLink()
}
\details{
\enumerate{
\item list all supprted organism
keggList("organism")
\item list genes, for one organism
keggList("dme")
\item convert ids (idtype)
keggConv("ncbi-proteinid", c("hsa:10458", "ece:Z5100"))
head(keggConv("eco", "ncbi-geneid")) # all entries
}

to_kegg_id
}
