% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{hash_string}
\title{Functions for HiSeq pipeline, utils for general data processing}
\usage{
hash_string(x)
}
\arguments{
\item{x}{string}
}
\description{
Functions for HiSeq pipeline, utils for general data processing
}
