% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_utils.R
\name{valid_featurecounts_input}
\alias{valid_featurecounts_input}
\title{valid_featurecounts_input
for function \code{import_featurecounts()}}
\usage{
valid_featurecounts_input(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
valid_featurecounts_input
for function \code{import_featurecounts()}
}
