% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{venn_plot}
\alias{venn_plot}
\title{deprecated
venn_plot Create venn plot using ggVennDiagram}
\usage{
venn_plot(..., names = NULL)
}
\arguments{
\item{...}{list, unique ids for groups}

\item{names}{character, the names for each group}
}
\description{
deprecated
venn_plot Create venn plot using ggVennDiagram
}
