% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot Create volcano plot}
\usage{
volcano_plot(
  data,
  x,
  y,
  labels = NULL,
  xintercept = c(0.5, 2),
  yintercept = 0.05,
  show_sig = TRUE,
  title = "Volcano plot"
)
}
\arguments{
\item{data}{A data frame for the full version of data}

\item{x}{character Name of the column, show on X-axis, eg: fold-change}

\item{y}{character Name of the column, show on Y-axis, eg: p-value, padj}

\item{labels}{characters Text labels on the plot}

\item{xintercept}{numeric Dot lines on x-axis default: c(0.5, 2)}

\item{yintercept}{numeric Dot lines on y-axis, default: 0.05}

\item{show_sig}{bool Show significant points}

\item{title}{character Title, default: "Volcano plot"}
}
\description{
volcano_plot Create volcano plot
}
