% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors}
\alias{colors}
\title{Functions for colors}
\usage{
colors(x)
}
\value{
string
}
\description{
A tmp function for documentation
}
\details{
This is an implementation of pick_colors, from ggplot2 default, or
RColorBrewer, fishualize

colors
}
