% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{compare_deseq}
\alias{compare_deseq}
\title{compare_deseq}
\usage{
compare_deseq(x, y, feature = "gene")
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{feature}{string default: gene}
}
\description{
compare_deseq
}
