% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq.R
\name{deseq}
\alias{deseq}
\title{Functions for DESeq2 down-stream analysis}
\usage{
deseq(dds, ...)
}
\arguments{
\item{dds}{DESeqDataSet}

\item{outdir}{character saving the results}

\item{strandness}{character could be "sens", "anti", default "sens"}

\item{fix_batch}{bool fix batch effect, from replicates, suffix; default: TRUE}

\item{shrink}{logical \code{shrink} LFC by \link{"apeglm", "ashr", "normal"},
default: TRUE}

\item{transform}{logical transform dds by \code{vst()}, \code{rlog()}, default: TRUE}

\item{overwrite}{bool overwrite exists file, default: FALSE}

\item{readable}{add \code{ENTREZID}, \code{SYMBOL}, based on gene_id, require: \code{genome}}

\item{genome}{character name of the organism, could be dm6, fruitfly}

\item{fc}{numeric cutoff for foldchange, default: 2, ignore foldchange}

\item{pvalue}{numeric cutoff for padj, default: 0.05, the main criteria}

\item{p_adjust}{bool use p-adjust value instead}

\item{cpu}{integer, number of CPU to run in parallel, default: 2}
}
\value{
DESeqResults, res (shrinked)
}
\description{
Reading files, directories, pipeline, ...
Processing data
Prepare for plot
Plotting
}
\details{
deseq
}
