% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{deseq_pair_stat_plot}
\alias{deseq_pair_stat_plot}
\title{deseq_pair_stat_plot}
\usage{
deseq_pair_stat_plot(x = NULL, y = NULL, .pd_rds = NULL)
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{.pd_rds}{string path to the file, saving read_deseq_pair() output}

\item{feature}{string default: gene}
}
\description{
deseq_pair_stat_plot
}
