% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_qc.R
\name{deseq_qc}
\alias{deseq_qc}
\title{Functions for DESeq2, dds quality control}
\usage{
deseq_qc(x, ...)
}
\arguments{
\item{x}{character file \code{norm_table.fix.csv}, or \code{DESeqDataSet}
also support the following data types: (be careful)
\code{data.frame}, \code{.rds}#'

generating the following plots:

deseq_qc_counts()
deseq_qc_mean_sd() # standard deviation
deseq_qc_top_gene()
deseq_qc_dist() # correlation
deseq_qc_pca()  # correlation
deseq_qc_ma()
deseq_qc_volcano()
deseq_qc_scatter()}

\item{outdir}{character saving the results}

\item{log2fc_limits}{numeric setting the range of log2FoldChange on plots
default: NULL, by \code{scales::breaks_extend()}}

\item{transform}{logical compute tranformed data, \code{vst()}, \code{rlog()}
default: FALSE}

\item{n_max}{integer number of genes to display, defualt: 16}

\item{fc}{numeric cutoff for foldchange, default: 1, ignore foldchange}

\item{pvalue}{numeric cutoff for padj, default: 0.1, the main criteria}

\item{p_adjust}{bool use p-adjust value instead}

\item{overwrite}{bool overwrite exists file, default: FALSE}
}
\value{
list of ggplots
}
\description{
deseq_qc_counts()
deseq_qc_mean_sd()
deseq_qc_top_gene()
deseq_qc_dist()
deseq_qc_pca()
deseq_qc_ma()
deseq_qc_volcano()
deseq_qc_scatter()
}
\details{
to-to:
\itemize{
\item simpify arguments by \code{...}
\item distance between samples (bam cor)
}
}
