% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falco.R
\name{falco_plot_base_content}
\alias{falco_plot_base_content}
\title{falco}
\usage{
falco_plot_base_content(x, mode = "line")

falco_plot_base_content(x, mode = "line")
}
\arguments{
\item{x}{output of falco_read_txt()}

\item{mode}{character line, bar}
}
\description{
read falco output
}
\details{
falco_plot_base_content
}
