% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go}
\alias{go}
\title{Functions for GO analysis}
\usage{
go(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{character Genes for GO analysis}

\item{organism}{character Name of the organism, eg: dm3, fruitfly}

\item{outdir}{string Path to the directory, saving go_group results
..., pass the following arguments from parent function

or else, set default values:
orgdb    = NULL
keytype  = NULL
fold_change = NULL
level    = 2
readable = TRUE
overwrite = FALSE}

\item{orgdb}{OrgDb from AnnotationDbi, overwrite \code{organism}}

\item{keytype}{character Name of the keytype for input, default: NULL}

\item{level}{int GO levels, default: 2}

\item{readable}{bool Args for groupGO function, convert to gene symbol

... pass
text_width
organism, and so on, to children functions

go_group(), keytype, level, readable
go_enrich(), keytype, readable, pval_cutoff, qval_cutoff
go_gsea(), keytype, pval_cutoff}
}
\description{
organize GO output
generate tables, plots
}
\details{
go_group
\enumerate{
\item csv files, results
\item barplot,
\item wego plot,
}

go_enrich
\enumerate{
\item csv files, results
\item barplot
\item wego plot,
\item dotplot
\item netplot
\item ...
}

go
}
