% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_plot.R
\name{go_emapplot}
\alias{go_emapplot}
\title{create plots}
\usage{
go_emapplot(x, ...)
}
\arguments{
\item{x}{object of enrichGO}

\item{..., }{support, fold_change, text_width,

orgdb
show_category
layout, kk, nicely,

updated: 2020-12-15, pairwise_termsim(ego)}
}
\description{
create plots
}
