% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrich.R
\name{go_enrich_plot}
\alias{go_enrich_plot}
\title{go_enrich_plot}
\usage{
go_enrich_plot(x, ...)
}
\arguments{
\item{...}{passing extra arguments
parent function: get_go_plots(),

fold_change
text_width}

\item{gene_list}{object of enrichGO}
}
\description{
go_enrich_plot
}
