% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_group.R
\name{go_group}
\alias{go_group}
\title{Functions for GO analysis}
\usage{
go_group(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{character Genes for GO analysis}

\item{organism}{character Name of the organism, eg: dm3, fruitfly}

\item{outdir}{string Path to the directory, saving go_group results
..., pass the following arguments from parent function

or else, set default values:
orgdb    = NULL
keytype  = NULL
level    = 2
readable = TRUE}

\item{orgdb}{OrgDb from AnnotationDbi, overwrite \code{organism}}

\item{keytype}{character Name of the keytype for input, default: NULL}

\item{level}{int GO levels, default: 2}

\item{readable}{bool Args for groupGO function, convert to gene symbol}
}
\description{
organize GO output
generate tables, plots
}
\details{
go_group
\enumerate{
\item csv files, results
\item barplot,
\item wego plot,
}

go_enrich
\enumerate{
\item csv files, results
\item barplot
\item wego plot,
\item dotplot
\item netplot
\item ...
}
}
