% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_plot.R
\name{go_gsea_plot}
\alias{go_gsea_plot}
\title{go_gsea_plot}
\usage{
go_gsea_plot(x, ...)
}
\arguments{
\item{x}{object of gseaResult}

\item{..., }{support, fold_change, text_width,
fold_change
text_width
show_category}
}
\description{
go_gsea_plot
}
