% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_pipe.R
\name{go_pipe}
\alias{go_pipe}
\title{run go analysis}
\usage{
go_pipe(x, organism, outdir, fold_change = NULL, feature = "gene")
}
\arguments{
\item{x}{string, deseq_dir or gene names}

\item{organism}{string scientific name, eg: Drosophila melanogaster}

\item{outdir}{string, where to save the results}

\item{fold_change}{numeric, with gene_names,}

\item{feature}{string, required for deseq_dir}
}
\description{
run go analysis
}
