% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_pipe.R
\name{go_report}
\alias{go_report}
\title{go_report}
\usage{
go_report(x, feature = "gene")
}
\arguments{
\item{x}{path to GO output or deseq_dir}

\item{feature}{string support for deseq_dir, default: gene}
}
\description{
go_report
}
