% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_enrich.R
\name{hiseq_enrich}
\alias{hiseq_enrich}
\title{Functions for enrich analysis}
\usage{
hiseq_enrich(x, ...)
}
\arguments{
\item{x}{character path to the rnaseq_rx directory}
}
\value{
results
}
\description{
Path/DB: GO, KEGG (DO, MSigDb, Reactome, MeSH)
\itemize{
\item enrich: over-representation analysis
\item GSEA: Gene Set Enrichment analysis
}
}
\details{
including:
\itemize{
\item enrich_go
\item enrich_kegg
\item gsea_go
\item gsea_kegg
\item group_go
...
}

Adapt the functions from clusterProfiler, enrichplot, ...

\code{enrichGO()}, \code{enrichKEGG()}, \code{gseGO}, \code{gseKEGG}, \code{groupGO}, ...
}
