% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_compare_cluster.R
\name{hiseq_prep_compare_go}
\alias{hiseq_prep_compare_go}
\title{functions for compareCluster
hiseq_prep_compare_go
prepare data for \code{compareCluster()} function}
\usage{
hiseq_prep_compare_go(x, ...)
}
\arguments{
\item{x}{character path to the rnaseq_rx, parsing the mut_name}

\item{outdir}{string path to the directory, saving results}

\item{fc}{numeric fold-change cutoff, default: 2}

\item{pvalue}{float cutoff for pvalue, default: 0.05}

\item{p_adjust}{logical, whether use adjusted pvalue, default: TRUE}

\item{shrink_method}{string how to shrink the fold-change results}
}
\value{
data.frame including columns: gene_id, group, sig, genome

arguments
}
\description{
functions for compareCluster
hiseq_prep_compare_go
prepare data for \code{compareCluster()} function
}
