% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiseq_summary.R
\name{hiseq_summary}
\alias{hiseq_summary}
\title{Functions for HiSeq summary}
\usage{
hiseq_summary(x, hiseq_type = "auto")
}
\arguments{
\item{x}{path to the ATACseq project}
}
\description{
Summary data in project_dir.
Processing data
Prepare for plot
Plotting
}
