% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{is_valid_sheet_df}
\alias{is_valid_sheet_df}
\title{is_valid_sheet_df}
\usage{
is_valid_sheet_df(df, verbose = FALSE)
}
\arguments{
\item{data.frame}{sample data}

\item{index_list}{list, named index sequence}
}
\description{
is_valid_sheet_df
}
