% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{kegg}
\alias{kegg}
\title{Functions for KEGG analysis}
\usage{
kegg(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{vector of gene list}

\item{organism}{name of genome, eg: Homo sapiens}

\item{...}{extra argument

fold_change, pval_cutoff, qval_cutoff}

\item{outdir}{string}
}
\description{
overrepresent analysis
GSEA analysis
}
\details{
enrich kegg
\enumerate{
\item csv files, results
\item barplot
\item wego plot,
\item dotplot
\item netplot
\item ...
}

gsea kegg
\enumerate{
\item csv files, results
\item barplot
\item dotplot
\item netplot
\item ...
}

run_kegg
}
