% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_enrich.R
\name{kegg_enrich}
\alias{kegg_enrich}
\title{Functions for KEGG analysis}
\usage{
kegg_enrich(gene_list, organism, ...)
}
\arguments{
\item{gene_list}{character Genes for KEGG analysis}

\item{organism}{character Name of the organism, eg: dm3, fruitfly}

\item{...}{pass arguments: , orgdb, keytype, pval_cutoff, qval_cutoff,
readable, with default values:
keytype = NULL
pval_cutoff = 0.9
qval_cutoff = 0.9
readable = TRUE}

\item{keytype}{character Name of the keytype for input, default: NULL}

\item{pval_cutoff}{float Cutoff for p-value, default: 0.9}

\item{qval_cutoff}{float Cutoff for q-value, default: 0.9}

\item{readable}{bool Args for enrichKEGG function, convert to gene symbol}
}
\description{
pval_cutoff, qval_cutoff, set to 0.9, in order to return
enrich results anyway, filter in downstream analysis
}
\details{
overrepresent analysis
GSEA analysis

enrich kegg
\enumerate{
\item csv files, results
\item barplot
\item wego plot,
\item dotplot
\item netplot
\item ...
}
}
