% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diy.R
\name{load_te_tissue}
\alias{load_te_tissue}
\title{load te tissue specific annotation}
\usage{
load_te_tissue(...)
}
\description{
Data from this publication:
1.Malone, C. D. et al. Specialized piRNA pathways act in germline and somatic tissues of the Drosophila ovary. Cell vol. 137 522–535 (2009). DOI: \href{https://doi.org/10.1016/j.cell.2009.03.040}{10.1016/j.cell.2009.03.040}
}
