% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{make_publish_plots}
\alias{make_publish_plots}
\title{make_publish_plots}
\usage{
make_publish_plots(x, outdir = NULL, to_pdf = TRUE)
}
\arguments{
\item{x}{character rnaseq_rx directory, a.vs.b}

\item{outdir}{character Path to the directory, saving the plots}

\item{save2pdf}{bool Save the plots in PDF file,}
}
\description{
make_publish_plots
}
