% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffbind.R
\name{prep_diffbind}
\alias{prep_diffbind}
\title{prep_diffbind}
\usage{
prep_diffbind(design, outdir)
}
\arguments{
\item{design}{data.frame or csv file, for the peak and bam files}

\item{outdir}{output dirs}
}
\description{
prep_diffbind
}
