% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{save_go_plot}
\alias{save_go_plot}
\title{save plots to outdir}
\usage{
save_go_plot(x, outdir, name = NULL)
}
\arguments{
\item{x}{list of go plots}

\item{outdir}{string, path to save the plots}

\item{name}{prefix for the plots}
}
\description{
save plots to outdir
}
