% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{save_go_table}
\alias{save_go_table}
\title{save_go_table}
\usage{
save_go_table(x, outdir, name = NULL)
}
\arguments{
\item{x}{list of go results}

\item{outdir}{string, path to save the plots}

\item{name}{prefix for the plots}
}
\description{
save_go_table
}
