% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_cmp.R
\name{stat_deseq_pair}
\alias{stat_deseq_pair}
\title{stat_deseq_pair}
\usage{
stat_deseq_pair(
  x = NULL,
  y = NULL,
  .pd_rds = NULL,
  keep_not_sig = TRUE,
  group = "all",
  ...
)
}
\arguments{
\item{x}{string to GroupA "A.vs.B"}

\item{y}{string to GroupB "A.vs.B"}

\item{.pd_rds}{string path to the file, saving read_deseq_pair() output}

\item{group}{character "all", "gene", "te", "piRC"}

\item{kepp_not_sig}{logical whether keep not sig changed genes, default: TRUE}
}
\description{
stat_deseq_pair
}
